// src/sanity/plugin/sanity-studio/src/types.ts
export type slugPrefix =
  | string
  | ((doc: SanityDocument, context: SlugContext) => Promise<string> | string)

export type slugSourceFn = (
  document: SanityDocument,
  context: SlugContext,
) => string | Promise<string>

export type slugOptions = Pick<SlugOptions, 'isUnique'> & {
  source?: string | Path | slugSourceFn
  prefix?: slugPrefix
  folder?: {
    canUnlock?: boolean
  }
  i18n?: {
    enabled?: boolean
    defaultLocaleId?: string
    localizeslug?: LocalizeslugFn
  }
  autoNavigate?: boolean
}

export type slugParams = Omit<
  SlugDefinition & FieldDefinitionBase,
  'type' | 'options' | 'name'
> & {
  name?: string
  options?: slugOptions
}